finalizado
asdasd

+++ src/main/java/entidades/EvaluacionesCalificadas.java
@@ -1,29 +1,22 @@
-/*
- * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
- */
 package entidades;
 
-/**
- *
- * @author renzo
- */
 public class EvaluacionesCalificadas {
+
     private String nombreEvaluacion;
     private String siglasEvaluacion;
-    private double peso;
+    private int peso;
     private String cronograma;
 
     public EvaluacionesCalificadas() {
     }
 
-    public EvaluacionesCalificadas(String nombreEvaluacion, String siglasEvaluacion, double peso, String cronograma) {
+    public EvaluacionesCalificadas(String nombreEvaluacion, String siglasEvaluacion, int peso, String cronograma) {
         this.nombreEvaluacion = nombreEvaluacion;
         this.siglasEvaluacion = siglasEvaluacion;
         this.peso = peso;
         this.cronograma = cronograma;
     }
-    
+
     public String getNombreEvaluacion() {
         return nombreEvaluacion;
     }
@@ -40,11 +33,11 @@ public class EvaluacionesCalificadas {
         this.siglasEvaluacion = siglasEvaluacion;
     }
 
-    public double getPeso() {
+    public int getPeso() {
         return peso;
     }
 
-    public void setPeso(double peso) {
+    public void setPeso(int peso) {
         this.peso = peso;
     }
 
@@ -55,8 +48,4 @@ public class EvaluacionesCalificadas {
     public void setCronograma(String cronograma) {
         this.cronograma = cronograma;
     }
-    
-    
-    
-    
 }
